refactor(web): separate account deletion ownership
