+++ apps/web/package.json
@@ -6,7 +6,7 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "preview": "vite preview",
-    "test": "vitest run src/appData/domain/index.test.ts src/appData/session/useWorkspaceSession*.test.tsx src/appData/sync/observation/syncLifecycleObservation.test.ts src/appData/sync/mediaUploads/*.test.ts src/appData/sync/engine/useSyncEngine.mediaUploadLifecycle.test.tsx src/localDb/core/core.migrations.test.ts src/localDb/reviews/reviews.test.ts src/localDb/progress/progress.test.ts src/localDb/reviews/reviewSchedule.test.ts src/localDb/cards/cards.test.ts src/localDb/feedback/feedback.test.ts src/localDb/mobileAppPromotion/mobileAppPromotion.test.ts src/localDb/mediaTransfers/mediaTransfers.test.ts src/media/*.test.ts src/feedback/**/*.test.ts src/screens/review/**/*.test.ts src/screens/review/**/*.test.tsx src/screens/review/components/card/*.test.tsx src/screens/review/reactions/lottie/*.test.ts src/screens/review/tests/controls/*.test.tsx src/screens/cards/form/*.test.tsx src/screens/settings/**/*.test.tsx src/screens/progress/*.render.test.tsx src/screens/invite/**/*.test.tsx src/screens/catalog/*.test.tsx src/screens/share/*.test.tsx src/appData/progress/source/progressSource.summarySeries.test.tsx src/appData/progress/source/progressSource.cache.test.tsx src/appData/progress/source/progressSource.reviewSchedule.test.tsx src/appData/progress/source/progressSource.lifecycle.test.tsx src/appData/progress/source/progressSource.streakLeaderboard.test.tsx src/appData/progress/invalidation/progressInvalidation.test.tsx src/progress/progressDates.test.ts src/api/authUrls.test.ts src/api/endpoints/*.test.ts src/apiContracts/*.test.ts src/api/transport/*.test.ts src/chat/streaming/liveStream.test.ts src/chat/attachments/*.test.ts src/chat/ChatPanel/tests/*.test.tsx src/chat/ChatPanel/tests/*/*.test.tsx src/chat/composer/*.test.ts src/chat/composer/*/*.test.ts src/accountDeletion.test.ts src/appError/AppErrorContext.test.tsx src/observability/instrument.test.ts",
+    "test": "vitest run src/appData/domain/index.test.ts src/appData/session/useWorkspaceSession*.test.tsx src/appData/sync/observation/syncLifecycleObservation.test.ts src/appData/sync/mediaUploads/*.test.ts src/appData/sync/engine/useSyncEngine.mediaUploadLifecycle.test.tsx src/localDb/core/core.migrations.test.ts src/localDb/reviews/reviews.test.ts src/localDb/progress/progress.test.ts src/localDb/reviews/reviewSchedule.test.ts src/localDb/cards/cards.test.ts src/localDb/feedback/feedback.test.ts src/localDb/mobileAppPromotion/mobileAppPromotion.test.ts src/localDb/mediaTransfers/mediaTransfers.test.ts src/media/*.test.ts src/feedback/**/*.test.ts src/screens/review/**/*.test.ts src/screens/review/**/*.test.tsx src/screens/review/components/card/*.test.tsx src/screens/review/reactions/lottie/*.test.ts src/screens/review/tests/controls/*.test.tsx src/screens/cards/form/*.test.tsx src/screens/settings/**/*.test.tsx src/screens/progress/*.render.test.tsx src/screens/invite/**/*.test.tsx src/screens/catalog/*.test.tsx src/screens/share/*.test.tsx src/appData/progress/source/progressSource.summarySeries.test.tsx src/appData/progress/source/progressSource.cache.test.tsx src/appData/progress/source/progressSource.reviewSchedule.test.tsx src/appData/progress/source/progressSource.lifecycle.test.tsx src/appData/progress/source/progressSource.streakLeaderboard.test.tsx src/appData/progress/invalidation/progressInvalidation.test.tsx src/progress/progressDates.test.ts src/api/authUrls.test.ts src/api/endpoints/*.test.ts src/apiContracts/*.test.ts src/api/transport/*.test.ts src/chat/streaming/liveStream.test.ts src/chat/attachments/*.test.ts src/chat/ChatPanel/tests/*.test.tsx src/chat/ChatPanel/tests/*/*.test.tsx src/chat/composer/*.test.ts src/chat/composer/*/*.test.ts src/accountDeletion/accountDeletionAttempt.test.ts src/appData/session/browserSessionRecovery.test.ts src/appError/AppErrorContext.test.tsx src/observability/instrument.test.ts",
     "test:e2e": "FLASHCARDS_E2E_TARGET=prod playwright test",
     "test:e2e:local": "node ./scripts/check-local-e2e-stack.mjs && FLASHCARDS_E2E_TARGET=local playwright test",
     "test:e2e:prod": "FLASHCARDS_E2E_TARGET=prod playwright test"
