+++ apps/web/src/appData/session/browserSessionRecovery.test.ts
@@ -2,35 +2,35 @@
 import "fake-indexeddb/auto";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import {
-  beginAccountDeletionRetryAttempt,
-  clearAllLocalBrowserData,
-  clearBrowserReauthRequired,
   hasAccountDeletionAttemptDispatched,
   isAccountDeletionPending,
   isAccountDeletionServerConfirmed,
-  isBrowserReauthRequired,
   loadAccountDeletionAttemptId,
   loadAccountDeletionCsrfToken,
   markAccountDeletionAttemptDispatched,
   markAccountDeletionServerConfirmed,
-  markBrowserReauthRequired,
-  runWithAccountDeletionLock,
   setAccountDeletionPending,
   storeAccountDeletionCsrfToken,
-} from "./accountDeletion";
-import { AI_CHAT_COMPOSER_SUGGESTIONS_STORAGE_KEY } from "./chat/preferences/AIChatPreferencesContext";
-import { INSTALLATION_ID_STORAGE_KEY } from "./clientIdentity";
-import { LOCALE_PREFERENCE_STORAGE_KEY } from "./i18n/runtime";
-import { loadCloudSettings, putCloudSettings } from "./localDb/sync/cloudSettings";
-import { clearWebSyncCache } from "./localDb/cache";
-import { SYNC_RESTORE_HISTORY_STORAGE_KEY } from "./appData/sync/restore/syncRestoreHistory";
-import type { CloudSettings } from "./types";
+} from "../../accountDeletion/accountDeletionAttempt";
+import { SYNC_RESTORE_HISTORY_STORAGE_KEY } from "../sync/restore/syncRestoreHistory";
+import { AI_CHAT_COMPOSER_SUGGESTIONS_STORAGE_KEY } from "../../chat/preferences/AIChatPreferencesContext";
+import { INSTALLATION_ID_STORAGE_KEY } from "../../clientIdentity";
+import { LOCALE_PREFERENCE_STORAGE_KEY } from "../../i18n/runtime";
+import { clearWebSyncCache } from "../../localDb/cache";
+import { loadCloudSettings, putCloudSettings } from "../../localDb/sync/cloudSettings";
+import type { CloudSettings } from "../../types";
+import {
+  clearAllLocalBrowserData,
+  clearBrowserReauthRequired,
+  isBrowserReauthRequired,
+  markBrowserReauthRequired,
+} from "./browserSessionRecovery";
 
 const observabilityMocks = vi.hoisted(() => ({
   addWebBreadcrumbMock: vi.fn(),
 }));
 
-vi.mock("./observability/webObservability", () => ({
+vi.mock("../../observability/webObservability", () => ({
   addWebBreadcrumb: observabilityMocks.addWebBreadcrumbMock,
 }));
 
@@ -45,27 +45,6 @@ const seededCloudSettings: CloudSettings = {
 };
 
 function ignoreIndexedDbOpenRecoveryFailure(): void {
-}
-
-function installSerialAccountDeletionLockMock(): void {
-  let previousRequest: Promise<void> = Promise.resolve();
-  Object.defineProperty(window.navigator, "locks", {
-    configurable: true,
-    value: {
-      request<Result>(
-        _name: string,
-        _options: Readonly<{ signal: AbortSignal }>,
-        action: () => Promise<Result>,
-      ): Promise<Result> {
-        const result = previousRequest.then(action);
-        previousRequest = result.then(
-          (): void => undefined,
-          (): void => undefined,
-        );
-        return result;
-      },
-    },
-  });
 }
 
 function createStorageMock(): Storage {
@@ -180,12 +159,11 @@ beforeEach(async () => {
 afterEach(async () => {
   window.localStorage.clear();
   clearBrowserReauthRequired();
-  Reflect.deleteProperty(window.navigator, "locks");
   vi.restoreAllMocks();
   await clearWebSyncCache();
 });
 
-describe("account deletion local cleanup helpers", () => {
+describe("browser session recovery", () => {
   it("completes cleanup when the database delete is blocked before succeeding", async () => {
     seedLocalBrowserState();
     await putCloudSettings(seededCloudSettings);
@@ -323,135 +301,6 @@ describe("account deletion local cleanup helpers", () => {
 
     expect(isAccountDeletionPending()).toBe(true);
     expect(loadAccountDeletionCsrfToken()).toBe("csrf-token");
-    expect(isAccountDeletionServerConfirmed()).toBe(true);
-    expect(hasAccountDeletionAttemptDispatched()).toBe(true);
-  });
-
-  it("does not infer server confirmation from a new pending deletion", () => {
-    markAccountDeletionServerConfirmed();
-    storeAccountDeletionCsrfToken("csrf-token");
-
-    setAccountDeletionPending(true);
-
-    expect(isAccountDeletionPending()).toBe(true);
-    expect(loadAccountDeletionCsrfToken()).toBe("csrf-token");
-    expect(isAccountDeletionServerConfirmed()).toBe(false);
-  });
-
-  it("resets and clears the dispatch claim with pending attempt lifecycle", () => {
-    setAccountDeletionPending(true);
-    const firstAttemptId = loadAccountDeletionAttemptId();
-    markAccountDeletionAttemptDispatched();
-
-    expect(firstAttemptId).not.toBeNull();
-    expect(hasAccountDeletionAttemptDispatched()).toBe(true);
-
-    setAccountDeletionPending(true);
-
-    expect(loadAccountDeletionAttemptId()).not.toBe(firstAttemptId);
-    expect(hasAccountDeletionAttemptDispatched()).toBe(false);
-
-    markAccountDeletionAttemptDispatched();
-    setAccountDeletionPending(false);
-
-    expect(loadAccountDeletionAttemptId()).toBeNull();
-    expect(hasAccountDeletionAttemptDispatched()).toBe(false);
-  });
-
-  it("keeps server confirmation when the following recovery checkpoint throws", () => {
-    setAccountDeletionPending(true);
-    const recoveryError = new Error("IndexedDB recovery required");
-
-    expect(() => {
-      markAccountDeletionServerConfirmed();
-      throw recoveryError;
-    }).toThrow(recoveryError);
-
-    expect(isAccountDeletionPending()).toBe(true);
-    expect(isAccountDeletionServerConfirmed()).toBe(true);
-  });
-
-  it("prevents automatic waiters from dispatching after an ordinary owner failure", async () => {
-    installSerialAccountDeletionLockMock();
-    setAccountDeletionPending(true);
-    let deleteRequestCount = 0;
-    const recoveryController = new AbortController();
-    const ordinaryDeleteError = new Error("Account deletion request failed");
-
-    const completePendingDeletion = (): Promise<void> => runWithAccountDeletionLock(
-      recoveryController.signal,
-      async (): Promise<void> => {
-        if (
-          isAccountDeletionPending() === false
-          || isAccountDeletionServerConfirmed()
-          || hasAccountDeletionAttemptDispatched()
-        ) {
-          return;
-        }
-
-        markAccountDeletionAttemptDispatched();
-        deleteRequestCount += 1;
-        throw ordinaryDeleteError;
-      },
-    );
-
-    const results = await Promise.allSettled([
-      completePendingDeletion(),
-      completePendingDeletion(),
-      completePendingDeletion(),
-    ]);
-
-    expect(deleteRequestCount).toBe(1);
-    expect(results[0]).toEqual({ status: "rejected", reason: ordinaryDeleteError });
-    expect(results[1]).toEqual({ status: "fulfilled", value: undefined });
-    expect(results[2]).toEqual({ status: "fulfilled", value: undefined });
-    expect(isAccountDeletionPending()).toBe(true);
-    expect(isAccountDeletionServerConfirmed()).toBe(false);
-    expect(hasAccountDeletionAttemptDispatched()).toBe(true);
-  });
-
-  it("collapses simultaneous explicit retries into one new dispatched attempt", async () => {
-    installSerialAccountDeletionLockMock();
-    setAccountDeletionPending(true);
-    markAccountDeletionAttemptDispatched();
-    let deleteRequestCount = 0;
-    const recoveryController = new AbortController();
-
-    const completePendingDeletion = (): Promise<void> => runWithAccountDeletionLock(
-      recoveryController.signal,
-      async (): Promise<void> => {
-        if (
-          isAccountDeletionPending() === false
-          || isAccountDeletionServerConfirmed()
-          || hasAccountDeletionAttemptDispatched()
-        ) {
-          return;
-        }
-
-        markAccountDeletionAttemptDispatched();
-        deleteRequestCount += 1;
-        markAccountDeletionServerConfirmed();
-      },
-    );
-    const retryPendingDeletion = async (): Promise<void> => {
-      const expectedAttemptId = loadAccountDeletionAttemptId();
-      if (expectedAttemptId === null) {
-        return;
-      }
-
-      const didBeginRetryAttempt = await runWithAccountDeletionLock(
-        recoveryController.signal,
-        async (): Promise<boolean> => beginAccountDeletionRetryAttempt(expectedAttemptId),
-      );
-      if (didBeginRetryAttempt) {
-        await completePendingDeletion();
-      }
-    };
-
-    await Promise.all([retryPendingDeletion(), retryPendingDeletion()]);
-
-    expect(deleteRequestCount).toBe(1);
-    expect(isAccountDeletionPending()).toBe(true);
     expect(isAccountDeletionServerConfirmed()).toBe(true);
     expect(hasAccountDeletionAttemptDispatched()).toBe(true);
   });
